% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Demonstrate.R
\name{Demonstrate}
\alias{Demonstrate}
\title{Demonstrate: Visualizing output from Winnow for the iPlant Collaborative Validate Workflow}
\usage{
Demonstrate(dir, make.AUC.plot = TRUE,
  AUC.plot.title = "Mean AUC By Population Structure and Heritability",
  make.MAE.plot = TRUE,
  MAE.plot.title = "Mean MAE By Population Structure and Heritability",
  herit.strings = list("_03_", "_04_", "_06_"), herit.values = list(0.3,
  0.4, 0.6), struct.strings = list("PheHasStruct", "PheNPStruct"),
  struct.values = list(TRUE, FALSE))
}
\arguments{
\item{dir}{A character string: input the directory containing the text files you would like to use.}

\item{make.AUC.plot}{Option to make a AUC by Population Structure and Heritability. Default set to TRUE}

\item{AUC.plot.title}{A Character String: title of the AUC plot.
Default set to "Mean AUC By Population Structure and Heritability"}

\item{make.MAE.plot}{Option to make MAE by Population Structure and Heritability plot. Default set to TRUE}

\item{MAE.plot.title}{A Character String: title of the MAE. Default set to
"Mean MAE By Population Structure and Heritability"}

\item{herit.strings}{Details the strings to look for in filenames for determining the heritability
coefficients. Corresponds to the list in herit.values}

\item{herit.values}{Sets the values for heritability. Default sets to 0.3, 0.4, 0.6.}

\item{struct.strings}{Details the strings to look for in filenames for determining population structure.
Corresponds to the list in _struct.values_. Example: list("PheHasStruct", "PheNPStruct")}

\item{struct.values}{The list of values for population structure. Generally Boolean (i.e. TRUE or FALSE).
Value order should correspond to the _struct.strings_ variable. Example (to go with previous): list(TRUE, FALSE)}
}
\description{
This function allows you to enter a directory including multiple Winnow output files and creating
visualizations fromt the output.
}
\examples{
demonstrate (dir="path")
}
\keyword{GWAS}
\keyword{demonstrate}
\keyword{graph}
\keyword{knowntruth}
\keyword{validate}

