% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Demonstrate2.R
\name{Demonstrate2}
\alias{Demonstrate2}
\title{Demonstrate2: Visualizing output from Winnow for the iPlant Collaborative Validate Workflow}
\usage{
Demonstrate2(dir, make.pos.plot = TRUE,
  pos.plot.title = "True Positives by False Positives",
  make.error.plot = TRUE, error.plot.title = "Plot of AUC by MAE",
  extra.plots = TRUE, AUC.axis.min = 0, AUC.axis.max = 1,
  MAE.axis.min = 0, MAE.axis.max = 2)
}
\arguments{
\item{dir}{A character string: input the directory containing the text files you would like to use.}

\item{make.pos.plot}{Option to make a true positive by false positive plot. Default set to TRUE}

\item{pos.plot.title}{A Character String: title of the True/False Positives plot.
Default set to "True Positives by False Positives"}

\item{make.error.plot}{Option to make an AUC by MAE plot. Default set to TRUE}

\item{error.plot.title}{A Character String: title of the AUC/MAE plot. Default set to
"Plot of AUC by MAE"}

\item{extra.plots}{Option to make two histogram files. One file is for false positives and the
other for true positives. Default set to TRUE}

\item{AUC.axis.min}{Sets the minimum value for the y-axis on the AUC by MAE plot. Default
set to '0'.}

\item{AUC.axis.max}{Sets the maximum value for the y-axis on the AUC by MAE plot. Default
set to '1.0'.}

\item{MAE.axis.min}{Sets the minimum value for the x-axis on the AUC by MAE plot. Default
set to '0'.}

\item{MAE.axis.max}{Sets the maximum value for the x-axis on the AUC by MAE plot. Default
set to '2.0'.}
}
\description{
This function allows you to enter a directory including multiple Winnow output files and creating
visualizations fromt the output.
}
\examples{
demonstrate2 (dir="path")
}
\keyword{GWAS}
\keyword{demonstrate}
\keyword{graph}
\keyword{validate}

