% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MPlot.R
\name{MPlot}
\alias{MPlot}
\title{Function to create a Manhattan Plot from GWAS output}
\usage{
MPlot(dir)
}
\arguments{
\item{dir}{A character string: input the directory containing the text files you would like to use.}
}
\description{
This function allows you to create a Manhattan Plot with the size indicating effect size.
All colored points are considered significant at a P-Value of 0.05.
}
\examples{
MPlot(dir="path")
}
\keyword{GWAS}
\keyword{manhattanplot}

